Code refactor for Dual Dan
Code refactor for Dual Dan
