+++ dual-dan_replication/Code/dual_dan_predict.py
@@ -3,7 +3,6 @@ from flip_gradient import flip_gradient
 from units_all import *
 import time
 from sklearn import metrics as mt
-import os
 import scipy.io as io
 
 opt_type = 'adam'
@@ -18,7 +17,6 @@ class_labels = 2
 domain_labels = 1
 hidden_dnn = 300
 max_gradient_norm = 1.0
-num_steps = 150
 
 today = time.strftime('%Y%m%d')
 hour = time.strftime('%h')
@@ -32,7 +30,7 @@ class DomainModel(object):
 
     def classifier(self, x, reuse=None, scope=None):
         with tf.variable_scope(scope, reuse=reuse):
-            w_dt_0 = weight_variable([2 * self.hidden_rnn, self.cl_n_first])
+            w_dt_0 = weight_variable([self.dense_output, self.cl_n_first])
             b_dt_0 = bias_variable([self.cl_n_first])
             dt_h_fc0 = tf.nn.relu(tf.matmul(x, w_dt_0) + b_dt_0)
 
@@ -48,11 +46,11 @@ class DomainModel(object):
 
     def generate_source_target(self, x, scope=None):
         with tf.variable_scope(scope):
-            weights_st = weight_variable([2 * self.hidden_rnn * time_steps, 2 * self.hidden_rnn])
-            biases_st = bias_variable([2 * self.hidden_rnn])
+            weights_st = weight_variable([2 * self.hidden_rnn * time_steps, self.dense_output])
+            biases_st = bias_variable([self.dense_output])
             if cell_type == 'lstm':
                 if num_layers > 1:
-                    # define rnn-cell with tensor_flow
+                    # define rnn cells with tensor_flow
                     # forward direction cell
                     fw_cell_st = tf.contrib.rnn.MultiRNNCell([rnn.LSTMCell(self.hidden_rnn)
                                                               for _ in range(num_layers)])
@@ -86,7 +84,7 @@ class DomainModel(object):
                     # backward direction cell
                     bw_cell_st = rnn.BasicRNNCell(self.hidden_rnn)
 
-            # get rnn-cell outputs
+            # get rnn cell output
             l_outputs_st, a_st, b_st = rnn.static_bidirectional_rnn(fw_cell_st, bw_cell_st,
                                                                     x, dtype=tf.float32)
             l_outputs_st = tf.transpose(tf.stack(l_outputs_st, axis=0), perm=[1, 0, 2])
@@ -94,15 +92,13 @@ class DomainModel(object):
             l_outputs_st = tf.reshape(l_outputs_st, [-1, 2 * self.hidden_rnn * time_steps])
 
             outputs_st = tf.nn.tanh(tf.matmul(l_outputs_st, weights_st) + biases_st)
-            lo_gits_st = tf.reshape(outputs_st, [-1, 2 * self.hidden_rnn])
+            lo_gits_st = tf.reshape(outputs_st, [-1, self.dense_output])
 
-            ab_st = tf.concat((a_st[1], b_st[1]), axis=1)
-
-            return lo_gits_st, ab_st
+            return lo_gits_st
 
     def discriminator_source_target(self, x, scope=None, reuse=None):
         with tf.variable_scope(scope, reuse=reuse):
-            w_dt_0 = weight_variable([2 * self.hidden_rnn, self.cl_n_first])
+            w_dt_0 = weight_variable([self.dense_output, self.cl_n_first])
             b_dt_0 = bias_variable([self.cl_n_first])
             dt_h_fc0 = tf.nn.relu(tf.matmul(x, w_dt_0) + b_dt_0)
 
@@ -150,16 +146,14 @@ class DomainModel(object):
 
         with tf.name_scope('generate_st'):
             x_s = tf.unstack(self.X_st, time_steps, 1)
-            lo_gits_s, h_s = self.generate_source_target(x_s, scope='generate_st')
+            lo_gits_s = self.generate_source_target(x_s, scope='generate_st')
 
         with tf.name_scope('generate_ts'):
             x_t = tf.unstack(self.X_ts, time_steps, 1)
-            lo_gits_t, h_t = self.generate_source_target(x_t, scope='generate_ts')
+            lo_gits_t = self.generate_source_target(x_t, scope='generate_ts')
 
         self.features_st = lo_gits_s
         self.features_ts = lo_gits_t
-        self.h_st = h_s
-        self.h_ts = h_t
 
         with tf.name_scope('classifier_st'):
             lo_gits_st = self.classifier(self.features_st, scope='classifier')
@@ -169,7 +163,7 @@ class DomainModel(object):
                                                                             labels=self.Y_st)
 
         with tf.name_scope('mode_collapse_source_target'):
-            # for source domain data
+            # for source data
             cross_s_sm_01 = tf.expand_dims(self.features_st, 1)
             cross_s_sm_02 = tf.expand_dims(cross_s_sm_01, 1)
 
@@ -177,18 +171,11 @@ class DomainModel(object):
             sum_kernel_st = cross_s_sm_01 - cross_s_sm_02
             distance_source = tf.squeeze(tf.matmul(sum_kernel_s, sum_kernel_st, transpose_b=True))
 
-            cross_hs_sm_01 = tf.expand_dims(self.h_st, 1)
-            cross_hs_sm_02 = tf.expand_dims(cross_hs_sm_01, 1)
-
-            sum_kernel_hs = cross_hs_sm_01 - cross_hs_sm_02
-            sum_kernel_h_st = cross_hs_sm_01 - cross_hs_sm_02
-            distance_h_source = tf.squeeze(tf.matmul(sum_kernel_hs, sum_kernel_h_st, transpose_b=True))
-
-            niu_source = tf.exp((-self.sigma_kernel * distance_h_source))
+            niu_source = tf.exp((-self.sigma_kernel * distance_source))
             exp_x_source = tf.where(tf.is_nan(niu_source), tf.ones(niu_source.shape), niu_source)
             self.mc_source = exp_x_source * distance_source
 
-            # for target domain data
+            # for target data
             cross_t_sm_01 = tf.expand_dims(self.features_ts, 1)
             cross_t_sm_02 = tf.expand_dims(cross_t_sm_01, 1)
 
@@ -196,14 +183,7 @@ class DomainModel(object):
             sum_kernel_tt = cross_t_sm_01 - cross_t_sm_02
             distance_target = tf.squeeze(tf.matmul(sum_kernel_t, sum_kernel_tt, transpose_b=True))
 
-            cross_ht_sm_01 = tf.expand_dims(self.h_ts, 1)
-            cross_ht_sm_02 = tf.expand_dims(cross_ht_sm_01, 1)
-
-            sum_kernel_ht = cross_ht_sm_01 - cross_ht_sm_02
-            sum_kernel_h_tt = cross_ht_sm_01 - cross_ht_sm_02
-            distance_h_target = tf.squeeze(tf.matmul(sum_kernel_ht, sum_kernel_h_tt, transpose_b=True))
-
-            niu_target = tf.exp((-self.sigma_kernel * distance_h_target))
+            niu_target = tf.exp((-self.sigma_kernel * distance_target))
             exp_x_target = tf.where(tf.is_nan(niu_target), tf.ones(niu_target.shape), niu_target)
 
             self.mc_target = exp_x_target * distance_target
@@ -298,7 +278,8 @@ _, _, target_test, target_test_labels = read_data(time_steps, sen_len, False)
 def train_and_evaluate():
     """helper to run the model with different training modes."""
     hidden_rnn = 128  # the size of the lstm hidden state corresponding to the best trained model
-    saved_dir = "./dual_dan_model/" + 'dual_dan_peg_png_hg/'  # the directory stores the best trained model obtained from running dual_dan_HG.py file.
+
+    saved_dir = "your_local_path/dual-dan_replication/Data/dual_dan_model/" + 'dual_dan_predict_model/'  # the directory stores the best trained model obtained from running dual_dan.py file
 
     model = DomainModel(hidden_rnn)
 
@@ -306,6 +287,7 @@ def train_and_evaluate():
     config.gpu_options.allow_growth = True
 
     with tf.Session(config=config) as sess:
+
         # check whether we have the model trained or not
         check_point = tf.train.get_checkpoint_state(saved_dir)
         if check_point and tf.train.checkpoint_exists(check_point.model_checkpoint_path):
@@ -320,7 +302,7 @@ def train_and_evaluate():
         gen_target_test_batch = batch_generator([target_test_values, target_test_labels], int(tbs // 2))
         t_size = int(target_test.shape[0] // (tbs // 2))
 
-        result_file = open('./dual_dan_model/' + 'dual_dan_peg_png_hg.txt', 'a+')
+        result_file = open('your_local_path/dual-dan_replication/Data/dual_dan_model/' + 'dual_dan_predict_model_result.txt', 'a+')
 
         # for target
         full_y_predict_train = np.array([])
@@ -357,11 +339,16 @@ def train_and_evaluate():
             fnr = float(fn) / (tp + fn)
 
         print('fpr: %.5f ; fnr: %.5f ; trg_test_acc: %.5f ; trg_test_pre: %.5f ; trg_test_f1: %.5f '
-          '; trg_test_re: %.5f ; trg_test_auc: %.5f' % (fpr, fnr, trg_test_acc, trg_test_pre,
-                                                        trg_test_f1, trg_test_re, trg_test_auc))
-        result_file.write('fpr: %.5f ; fnr: %.5f ; trg_test_acc: %.5f ; trg_test_pre: %.5f ; trg_test_f1: %.5f '
-          '; trg_test_re: %.5f ; trg_test_auc: %.5f \n' % (fpr, fnr, trg_test_acc, trg_test_pre,
-                                                        trg_test_f1, trg_test_re, trg_test_auc))
+              '; trg_test_re: %.5f ; trg_test_auc: %.5f' % (fpr, fnr, trg_test_acc, trg_test_pre, trg_test_f1,
+                                                            trg_test_re, trg_test_auc))
+
+        result_file.write("fpr: %.5f; " % fpr)
+        result_file.write("fnr: %.5f; " % fnr)
+        result_file.write("trg_test_acc: %.5f; " % trg_test_acc)
+        result_file.write("trg_test_pre: %.5f; " % trg_test_pre)
+        result_file.write("trg_test_f1: %.5f; " % trg_test_f1)
+        result_file.write("trg_test_re: %.5f; " % trg_test_re)
+        result_file.write("trg_test_auc: %.5f\n" % trg_test_auc)
 
 
 print('dual domain adaptation testing')
