+++ dual-dan_replication/Data/dan_model/infor.txt
