+++ dual-dan_replication/Data/dual_dan_model/dual_dan_predict_model/infor.txt
