+++ dual-dan_replication/Data/dual_dan_model_save/Readme
@@ -1,0 +1,1 @@
+The Dual-DDAN method trained models with different parameters will be saved in the "model" folder
