--- dual-dan_replication/data_sets_numberized/infor.txt
@@ -1,1 +1,0 @@
-
