--- dual-dan_replication/dual_dan_model/dual_dan_peg_png/infor.txt
@@ -1,1 +1,0 @@
-Please use the link https://app.box.com/s/tdz8n5o9qq6hgyz0pbaefcsp9rxnvf36 for downloading the trained dual-dan model.
